Refactor circle diagram initialization, move CircleDiagramComponent to separate header
